Do not clobber on unresolved jump.
There is nothing to clobber actually, since we don't know where the jump
leads, we will have no successor.

Actually, in the presence of unresolved jumps everything is incorrect,
the graph itself, the SSA form, so there is nothing can be done.

In any case, removing a special treatment of SSA form simplifies an algorithm.