Revise shingled disasm; updated bap & simplicity
This update shinks the actual disassembler and reduces
the overall size of the code to be simpler. In addition,
it reflects updates in BAP, where instead of using the
library lifter directly, it abstracts the lifter into
a module interface and selects an experimental lifter
that is on another branch. Some other notes are that
dependencies were reduced and some minor changes in
pre ocamlbuild process to correct non repeatability.