test: changed the matching engine
Previously, the matching required that the expected set is a proper
subset of the program output. Now, the matching requires that every
regular expression specified in the expected corpora matches with
at least one substring in the program output, that is not matched by
other expectation, i.e., each expectation gets its own prize. This
is a maximal matching problem, that is solved using Ford-Fulkerson
algorithm.