solver: use original call as a proof term
So far we've used the artificially injected definition term
as a proof. That was robust, but somewhat inconvenient, as
this required us to store the generated output, since in
the default BIR output there is no this nodes. Now, we use
the original call term as a term proof. Not sure, though,
that it is entirely good idea, so I may fall back from this
decision at some point.