Quarantine: revert previous, fix propagation
So first of all this reverts previous commit, where I tried to tackle
loops by just no jumping into the already evaluated code. This was a
bogus idea, that breaks saluki test case. We need something more
sophisticated.

As a compensation for the inconvenience, this commit will fix a bug,
where we didn't propagate taint in statements of the form `x := x`.