quarantine: added loop escaping
We will escape loops (backtrack out of them), if we made more than
a fixed amount of iterations. Currently it is a constant, but I will
make a parameter soon.

We detect a loop using the following criterion -- the jump doesn't cross
the subroutine boundaries and we have already visited the target. Note,
this patch will severely increase the coverage, but will drastically
decrease the performance. I will investigate if it possible to optimize
it.