This PR removes quarantine from bap-plugins
The plugin is moved into BAP main repository. It gave birth for two
packages:

  - bap-microx - a microexecution library;
  - bap-taint-propagator - a plugin that registers a pass that
    propagates taints

Now instead of saying `--quarantine` one should say
`--propagate-taint`.

The installation is also much easier. As you may just install it
as `opam install bap-taint-propagator`, or just as `opam install bap`,
as it is part of bap meta package.