Saluki annotate program (#16)
* Saluki: annotate program with comments

This is a hacky solution for integration with IDA, that will transform
solution into comment attributes. Each term, that participate in some
model, will be annotated with a comment of the following syntax:

       <sat>:<defn>_<rule>:<patt>

where
   <sat> ::= "SAT" | "UNS"   :: whether model was satisfied
   <defn>,<rule> ::= string  :: names of defn and rule
   <patt> ::= string         :: pattern

* Joined comments by prefix and changed format

If a term proves a pattern for the same definition (i.e., if it proves
several models of the same definition), then it will be printed only
once. Also the new comment format is:

      <SAT>:<defn>_<rule>">" <pat>

as it looks more readable.