categorize specifications
This is a workaround to shorten the messages outputted into IDA.  Since,
definition name must be unique we can just choose a short name for
it. Instead, we will prefix each definition with the name of rule
category to which it applies and separate it from the rest of the rule
with a '/' symbol. Only the category name will passed to IDA. In case if
there is no category specified, then the full definition will be
passed. Again, this is a workaround, later we will get rid of sat/uns,
defn/rule, uniq defn requirement, and will only output violations.