Saluki annotate program (#17)
* Saluki: annotate program with comments

This is a hacky solution for integration with IDA, that will transform
solution into comment attributes. Each term, that participate in some
model, will be annotated with a comment of the following syntax:

       <sat>:<defn>_<rule>:<patt>

where
   <sat> ::= "SAT" | "UNS"   :: whether model was satisfied
   <defn>,<rule> ::= string  :: names of defn and rule
   <patt> ::= string         :: pattern

* Joined comments by prefix and changed format

If a term proves a pattern for the same definition (i.e., if it proves
several models of the same definition), then it will be printed only
once. Also the new comment format is:

      <SAT>:<defn>_<rule>">" <pat>

as it looks more readable.

* changed term attribution

now we use two distinct attributes:

 - saluki-sat: for satisfied rules
 - saluki-uns: for unsatisfied rules

attribute contains a string with a name of definition

* categorize specifications

This is a workaround to shorten the messages outputted into IDA.  Since,
definition name must be unique we can just choose a short name for
it. Instead, we will prefix each definition with the name of rule
category to which it applies and separate it from the rest of the rule
with a '/' symbol. Only the category name will passed to IDA. In case if
there is no category specified, then the full definition will be
passed. Again, this is a workaround, later we will get rid of sat/uns,
defn/rule, uniq defn requirement, and will only output violations.