A better commandline interface
It is better, than nothing actually, as there wasn't any interface
before.

The interface now allows to specify concrete properties to be checked.
If not specified, then all known properties are checked. The properties
are specified as a substring, i.e., the specified name occurs as a
substring of a property, then it will be checked. For example,
```
bap --saluki-check=ERR
```
will check all properties in category ERR

```
bap --saluki-check=malloc
```
will check all properties, that containt `malloc` in their name.

To specify multiple properties, separate it with a comma, e.g.,
```
bap --saluki-check=ERR,TAINT,BYPASS
```

This PR also disables a printout of the models (aka finded proofs),
they can be enabled with `--saluki-print-models` option.