first prototype of beagle
Idea
====

On a high level, Beagle is tracking everything that smells to him. If
something smells, then it will grab the track and follow, until the
trail temperature is above the given threshold, in that case it will
grab the target and mark it as his prey.

The temperature is estimated with the following recurrences:
   $$$
   t = t * (1 - alpha) + alpha * phi,
   $$$

   where $t$ - is a temperature, between 0 and 1,
               is our confidence,
               that whatever we're sniffing is our target;
         $alpha$ - sensitivity parameter (defines how fast
               the trail is loosing the temperature);
         $phi$ - a confidence that something smells
                (again between 0 and 1).
                Defines how fast we will hunt the prey.
                Currently, it can be either 0 if something doesn't
               smell, or 1-1e4 if it smells.

Implementation
==============

Beagle uses microexecution, to run each subroutine. Every time a word is
seen by the interpreter the beagle is targeted on it.

Current status
=============

Currently all the parameters are hardcoded, and many parameters should
be estimated, not specified by a user. The only predicate, that defines
that something smells is whether a character is_ascii or is_null.

Usage
=====

1. Build and install with

   make

2. Run from IDA with `Shift-S` and specify option `--beagle`
   and the attribute name `beagle-prey`

3. Use the text search (Alt-T) to navigate through all preys.